KCL-6892 Move createTemplateForCustomElement function to node.ts file
